Move WMS and WFS samples to data category
